requirements.txt
feedparser==6.0.8
line-bot-sdk==1.20.0
python-dateutil==2.8.2
pytz==2021.3
